refactor linking into its own file
